#26602 Core v0.15 Explicit path cache synchronization
Prior to a context being loaded up in the main window, the path cache is
now synced. This is to ensure that the local caches (and folders) are up
to date. In particular, this covers the use case where an object has
been renamed, folders have been unregistered and new folders have been
created. If the local path cache is out of date at this point, it will
be unaware of the change and still contain the cache old entries.

Uses new core API method tk.synchronize_filesystem_structure() and
therefore contains a constraint where it needs the v0.15.x core.

Closes #39.