For #31249, overwriting scene files
If we let the model refresh it's state before clicking save, clicking it will invoke _generate_path and the cached value will be read.

If we don't wait for the background task to finish indexing the files, then we will get partial results and an incorrect version number.

This is because when the FileModel requests for the cache file versions it doesn't honour the clean_only flag that _generate_path to it. Because of this, the file model actually returns dirty cache entries. Honouring the flag when communicating with the cache will return None when the model is still being populated and the _generate_path will compute the value itself on the spot instead of waiting for the background task.

Closes https://github.com/shotgunsoftware/tk-multi-workfiles2/pull/3