#38761 Tree view sorted alphabetically
The left hand side tree view is now sorted alphabetically. Previously, it would just display the raw data as it was being sent from the shotgun model. The shotgun model does not imply any ordering of the data, this is left to the client to handle (since it typically varies between use cases and sometimes, like in the case of the loader, you want to do a bunch of post processing of the raw model data before you have access to the display name which you want to order by).

Closes #33