#39903 Workaround for Hiero not updating save menus.
Hiero apparently has a bug whereby when using project.saveAs() to write a file
path, it does not update its own File > Save menus. This change adds a
workaround that takes advantage of Hiero providing access to the QAction object
for all of its menu items. We get a handle on the objects and set their text
directly to match what Hiero should display.

Closes #40.