SG-15693 add change context mode (#98)
Co-authored-by: Jean-François Boismenu <jfboismenu@users.noreply.github.com>
Co-authored-by: Jean-François Boismenu <jean.francois.boismenu@autodesk.com>

Adds a context change mode that means the open and save app options are replaced by a context change app.
In context change mode it contains the entity/context navigation windows, but no file select. Once a context has been picked you can press change context.

Tasks in the My tasks can be view can be double clicked to change context, when in context change mode.

Removed a line specifically blocking Mari, and added a Mari scene hook, which doesn't actually contain any scene operation logic, but satisfies workfiles' need to find an engine hook.

Refactoring:
Previously, the open and save dialog derive from the same FileFormBase class, but that class makes no assumption about the UI, because the save and open dialogs have different UIs.
However this meant a bunch of duplicated code in the derived classes, because both UI's shared common functionality.
Instead of opting to make context change a "mode" of the open dialog, I decided to make it it's own class derived from the FileFormBase and to cut down on the duplication across all three classes I moved a bunch of methods over to the base class. However this means that the base class now has some expectation of what will be in the UI.